+++ arch/arm/core/cortex_m/CMakeLists.txt
@@ -15,6 +15,7 @@ zephyr_library_sources(
   swap.c
   swap_helper.S
   irq_manage.c
+  prep_c.c
   )
 
 zephyr_library_sources_ifdef(CONFIG_GEN_SW_ISR_TABLE isr_wrapper.S)
